Improve message from QueryException
I hit an invalid query doing some goofy joins and had a hard time
figuring out what I did wrong in my query. It would have been much
easier to figure out the problem if the error contained the problem
query, especially when we went to the trouble of collecting it.