Add SchemaDialect::describeForeignKeys()
Another step of #18123. TableSchema requires columns to be present for
foreignkeys to reference them. This requires reflecting columns in order
to get foreign key data.

I'm planning on doing more refactoring on SchemaDialect to remove the
ephemeral `TableSchema` and to avoid reflecting columns for indexes and
foreign keys once all the new reflection methods are complete.